+++ apps/examples/src/examples/lib/patched.tsx
@@ -1,7 +1,12 @@
 import { patchMaterial } from "@material-composer/patch-material"
 import { MaterialNode } from "@react-three/fiber"
-import { DependencyList, useLayoutEffect, useMemo } from "react"
-import { IUniform, Material, MeshStandardMaterial } from "three"
+import {
+  IUniform,
+  Material,
+  MeshPhysicalMaterial,
+  MeshStandardMaterial
+} from "three"
+import { useManagedInstance } from "./useManagedInstance"
 
 type Constructor<T> = new (...args: any[]) => T
 
@@ -24,23 +29,20 @@ const makePatchedMaterialComponent = <
   uniforms,
   ...props
 }: PatchedMaterialProps<C>) => {
-  const material = useManagedInstance(() => {
-    const material = new ctor(...(args as any))
-    return patchMaterial(material, { vertexShader, fragmentShader, uniforms })
-  }, [vertexShader, fragmentShader, uniforms])
+  const material = useManagedInstance(
+    () =>
+      patchMaterial(new ctor(...(args as any)), {
+        vertexShader,
+        fragmentShader,
+        uniforms
+      }),
+    [vertexShader, fragmentShader, uniforms]
+  )
 
   return <primitive object={material} {...props} />
 }
 
 export const patched = {
-  MeshStandardMaterial: makePatchedMaterialComponent(MeshStandardMaterial)
-}
-
-const useManagedInstance = <T extends any>(
-  factory: () => T,
-  deps: DependencyList = []
-) => {
-  const instance = useMemo(() => factory(), deps)
-  useLayoutEffect(() => () => (instance as any).dispose?.(), [instance])
-  return instance
+  MeshStandardMaterial: makePatchedMaterialComponent(MeshStandardMaterial),
+  MeshPhysicalMaterial: makePatchedMaterialComponent(MeshPhysicalMaterial)
 }
